Workaround for bug 6609605: Hindi rendering regression in WebView
This patch changes the font data so that all anchor points are in
absolute coordinates, rather than relative to outline points. With
the workaround, we don't trigger the bug in the HarfBuzz/Skia
integration that causes incorrect outline point data to be read.

The font was prepared with FontForge. There's some risk of other
problems, but I tested the font fairly carefully and it looks fine.
(With a version prepared in TTX, the substitution of the different
width forms of the ivowelsigndeva got broken, likely due to a TTX
bug, as this seems to happen with roundtrip of unedited original).

Change-Id: I3f6fffce91c3ea0695e6005b051bcf4e9b6f6361